:label: [rename] view 밑의 파일들 stage 패키지로 묶음
