+++ lib/ui/view/stage/popo_play_view.dart
@@ -7,7 +7,7 @@ import 'package:pocket_pose/data/entity/socket_request/send_skeleton_request.dar
 import 'package:pocket_pose/data/remote/provider/socket_stage_provider_impl.dart';
 import 'package:pocket_pose/domain/entity/stage_player_list_item.dart';
 import 'package:pocket_pose/domain/entity/stage_skeleton_pose_landmark.dart';
-import 'package:pocket_pose/ui/view/ml_kit_camera_view.dart';
+import 'package:pocket_pose/ui/view/stage/ml_kit_camera_view.dart';
 import 'package:provider/provider.dart';
 
 enum StagePlayScore { bad, good, great, excellent, perfect, none }
